refactor: Refactor index.js, separate login in login.js
